move to separate package, reset registry for tests
