+++ core/src/main/java/com/linecorp/armeria/client/endpoint/properties/FileWatcherRegistry.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.client.endpoint;
+package com.linecorp.armeria.client.endpoint.properties;
 
 import static java.nio.file.StandardWatchEventKinds.ENTRY_CREATE;
 import static java.nio.file.StandardWatchEventKinds.ENTRY_DELETE;
@@ -32,7 +32,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 import com.google.common.annotations.VisibleForTesting;
 
-import com.linecorp.armeria.client.endpoint.FileWatcherRunnable.FileWatchEvent;
+import com.linecorp.armeria.client.endpoint.properties.FileWatcherRunnable.FileWatchEvent;
 
 /**
  * Wraps a {@link WatchService} and allows paths to be registered.
