+++ core/src/test/java/com/linecorp/armeria/client/endpoint/properties/RestartableThreadTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.client.endpoint;
+package com.linecorp.armeria.client.endpoint.properties;
 
 import static org.assertj.core.api.Assertions.assertThat;
